set up to work with vercel
