+++ BookNest/.env
@@ -1,0 +1,2 @@
+SESSION_SECRET=booknest-secret-key-VawpIVC4xS
+REDIS_URL=rediss://default:AYMDAAIjcDE4YTE0MzYxNTQ3ZDg0YjAyYThiYTA3NTNiMmI1MDE4NXAxMA@heroic-eel-33539.upstash.io:6379
