+++ BookNest/package.json
@@ -3,6 +3,7 @@
   "version": "1.0.0",
   "description": "A simple but practical and scalable webapp for browsing your favorite books.",
   "type": "module",
+  "main": "./api/index.js",
   "engines": {
     "node": "22.*"
   },
@@ -11,18 +12,23 @@
     "start": "node api/index.js",
     "build": "echo 'no build step required'"
   },
-  "author": "CSE4214-group5",
+  "author": "",
   "license": "ISC",
   "dependencies": {
     "@vercel/node": "^3.2.24",
     "bcryptjs": "^2.4.3",
+    "connect-redis": "^7.1.1",
     "daemon": "^1.1.0",
+    "dotenv": "^16.4.5",
     "ejs": "^3.1.10",
     "express": "^4.21.1",
     "express-session": "^1.18.1",
+    "ioredis": "^5.4.1",
     "jest": "^29.7.0",
+    "jsonwebtoken": "^9.0.2",
     "multer": "^1.4.5-lts.1",
     "nodemon": "^3.1.7",
+    "redis": "^4.7.0",
     "request": "^2.88.2",
     "sequelize": "^6.37.5",
     "sequelize-cli": "^6.6.2",
@@ -37,16 +43,5 @@
     "@types/express-session": "^1.18.0",
     "@types/multer": "^1.4.12",
     "connect-flash": "^0.1.1"
-  },
-  "main": "api/index.js",
-  "directories": {
-    "test": "test"
-  },
-  "keywords": [
-    "bookstore",
-    "library",
-    "ecommerce",
-    "retail",
-    "webapp"
-  ]
+  }
 }
