+++ BookNest/public/stylesheet.css
@@ -311,13 +311,6 @@ img.a {
   float: left;
 }
 
-h4 {
-  font-family: 'Comic Sans MS', Comic, serif;
-  padding: 25px;
-  font-size: 25px;
-  text-align: center;
-}
-
 .slider {
   width: 90%;
   height: 40vh;
